# perf: memoize db_engine_spec in database