--- a/superset-frontend/cypress-base/cypress/integration/dashboard/load.test.ts
+++ b/superset-frontend/cypress-base/cypress/integration/dashboard/load.test.ts
@@ -44,9 +44,7 @@ describe('Dashboard load', () => {
 
   it('should load in edit/standalone mode', () => {
     cy.visit(`${WORLD_HEALTH_DASHBOARD}?edit=true&standalone=true`);
-    cy.get('[data-test="discard-changes-button"]', { timeout: 10000 }).should(
-      'be.visible',
-    );
+    cy.get('[data-test="discard-changes-button"]').should('be.visible');
     cy.get('#app-menu').should('not.exist');
   });