### SUMMARY
A recent PR #14547 introduced a performance regression causing dataset metadata fetching to become very slow for datasets with large numbers of columns. I originally thought the type regexes were the problem, but when researching the problem more closely it turns out that just referencing `self.table.database.db_engine_spec` in a `TableColumn` instance cost ~6ms on my local machine. Multiply that by 1000 columns ~= 6000 ms. To get around this I added memoization to the semi-expensive regex, but also added memoizing for `Database.db_engine_spec`. This should also speed up query rendering a bit, as there was similar logic there.

### BEFORE #14547 (pre-regression)
For the World Bank dataset (328 cols), fetching the data took slightly less than 180ms before on my local machine (including the unnecessary 20 ms redirect):
![image](https://user-images.githubusercontent.com/33317356/118236149-665aef00-b49e-11eb-872f-7f4cbfc637d9.png)

### CURRENT (master)
For the same dataset, retrieval of data now takes ~10s!
![image](https://user-images.githubusercontent.com/33317356/118236606-f00abc80-b49e-11eb-9bc7-ab907f8b9084.png)

### AFTER
Retrieval is now slightly quicker than originally (including no redirect):
![image](https://user-images.githubusercontent.com/33317356/118241602-0582e500-b4a5-11eb-9ca8-ff02d02d8c6a.png)

### TEST PLAN
<!--- What steps should be taken to verify the changes -->

### ADDITIONAL INFORMATION
<!--- Check any relevant boxes with "x" -->
<!--- HINT: Include "Fixes #nnn" if you are fixing an existing issue -->
- [ ] Has associated issue:
- [ ] Changes UI
- [ ] Includes DB Migration (follow approval process in [SIP-59](https://github.com/apache/superset/issues/13351))
  - [ ] Migration is atomic, supports rollback & is backwards-compatible
  - [ ] Confirm DB migration upgrade and downgrade tested
  - [ ] Runtime estimates and downtime expectations provided
- [ ] Introduces new feature or API
- [ ] Removes existing feature or API